fix: start cooldown in new extruder plan instead of old to combat snowballing effect (CURA-1731)
snowballing: when we start cooling T0 before we require the temp of T1 to be reached, due to small discrepancy between the computed and actual heatup times, the time it takes to actually reach the required temp of T1 - while computed to be zero - causes T0 to drop in temperature in that time. This in turn will mean it will also take longer to heat up again, creating the same problem in the next extruder switch from T1 to T0, thereby snowballing into layer switches taking longer and longer.