Move testing of combing_retract_before_outer_wall down into addTravel().
So now the boolean that is passed to addPolygon() and addPolygonByOptimizer()
is less specific, it simply states whether the polygon is an outer wall or
not. It's only when we get into addTravel() that we check whether
combing_retract_before_outer_wall is true when we are moving to an outer wall.