fix: add initial travel move to previous layer (CURA-3339)
The initial travel move was a direct travel move which didn't use combing (because the previous position was unknown)
Now this travel move is corrected, but the Comb of the previous layer is used.

Note that this travel move is added after processMinimalLayerTime has been called, which results in slightly off fan speeds and layer slowdown for small layers, because the initial travel move was direct instead of combed.