In spiralizeWallSlice() increase max_dist2 to improve performance on shallow slopes.
When the slope of the surface becomes shallow, the wall lines get further apart
and once the lines are spaced further than a line width apart the object
becomes essentially unprintable in spiral mode as the layers will have no
support below.

So when looking for a point on the last layer that matches the current point
of interest we didn't bother to look for points further than a line width away.
However, if you have a shallow sloped surface that produces lines that are
just over a line width apart or maybe they wander back and forth and cross
that threshold more than once, the resulting line has some points interpolated
and some not which looks really ragged. Doubling the distance threshold results
in smoother lines as fewer points are not interpolated. Of course, it doesn't
make the object more printable as a spiral.