LayerPlan::writeGCode() now uses "virtual prime tower" for multi-material spiralized prints.
Before, the post extruder change prime was occurring after the nozzle had been moved back
to the print so it was leaving a big blob on the print. Now, it does the prime into thin
air before moving back to the model. The hope is that it is much easier to clean a bit of
extrude away from the outside of the print compared to removing a firmly attached blob.