Try to avoid outputting outer wall insets when processing a hole.
It's possible that the number of insets around a hole can be less than the number of
insets inside the part's outer wall. This commit tries to avoid outputting an outer wall
inset when processing a hole that doesn't have all of it's insets. It's not the end of the
world if the outer wall inset is output, it will just take a bit longer to print.