Only generate lines for ironing if ironing is enabled
Right now this doesn't matter, since we only generate a top surface for ironing. However if we need the top surface for something else in the future and we generate it anyway for that feature, we will automatically enable sanding as well. So also check for the setting here.

Contributes to issue CURA-3903.