Add PrintFeatureType::PurgeMove and compensate travel speed for purge moves
CURA-3094

The prime power purge move acts the same as normal extrusion travels.
This move is way too fast for extruding a large amount of material,
which, in this case, is the purge. We add a new PrintFeatureType
"PurgeMove" so we can identify a purge move and compensate its travel
speed if the extrusion amount is way too much.