Don't convert ConstPolygonRef to Polygons before performing intersections
That's just wasteful copying of data. Just directly use the ConstPolygonRef. One flaw of this commit is that it does need to do a copy when actually passing the polygon data to Clipper, but that is unavoidable (and hopefully can be optimised out by a clever compiler).

Contributes to issue CURA-3862.