Improve PathOrderOptimizer::getClosestPointInPolygon() handling of points on concave curves.
The function used to adjust the computed distance of each point from prev_point by a
factor based on the angle between the line segments that pass through the point. It still
does this but with some differences. Now, it only adjusts the distance when the point
is on a concave curve and the amount of adjustment is now proportional to not only the
amount of curvature but also to the un-adjusted distance. The further away the point is
from prev_point, the more the computed distance is reduced when it lies on a concave curve.