Now getClosestPointInPolygon() will apply corner preference when seam type is closest.
This currently uses an algorithm very similar to the original implementation in that the
amount the corner point is shifted is independent of how far away the point is. Currently,
a 90 deg corner looks as if it is 5mm closer than another point that is the same distance
away but is located on a relatively straight region of the outline. Let's see how this works
out.