Make the LineOrderOptimizer prefer to start at ends of a sequence of connected lines.
Previously, when processing zigzag patterns of lines, it was possible for the lines to be
output as disconnected pairs of lines (a zig and a zag together). This commit adds code that
determines the vertices that are at the end of a sequence of lines and it selects the end
vertex that is closest to the current point as the next to be output.

This still doesn't guarantee that the lines are output optimally but it does at least appear
to stop the disconnected line pairs problem.