When calculating comb travel distance, don't fail on meeting an unavoidable obstacle.
If the combing did find an unavoidable obstacle it would fall back to using the direct
distance which is the worst thing it could do. Now it will continue to calculate the full
combing distance and (hopefully) some other point will be closer and chosen to be the start
of the next chain.