When optimizing comb path, remove groups of points when it won't clash with the boundary.
Sometimes, the combing algorithm will produce a maximal length path that hugs the part
boundary when it could, in fact, be shorter and still be inside the part. For example,
Imagine a part shaped like the outline of a hand. The combing algorithm is quite happy to
create a path that goes up and down the fingers when travelling from one side of the palm
to the other when, perhaps, it could cut across the base of each finger.

A real optimizer would detect the excess travel and trim it to the minimum required by
"pulling tight" the path. This commit tries to achieve a similar result but with no claims
to being optimal. It's relatively cheap (computationally) and cheerful (can produce good
results).

What it does is look back at the sequence of points that have been inserted into the comb
path and see if they can be removed ("short-circuited") without causing the path to cross
the comb boundary. So paths that hug a concave boundary will be shortened. They can even be
shortened when they go around holes but that is not guaranteed to occur.