apply initial layer flow to initial model layer.
The initial layer height is also applied to the initial model layer, so the flow should be applied to the same thing.
Moreover, the first raft layer has only one line width and fill percentage, so you can adjust those if you want to adjust the first raft layer flow.