New algorithm - almost right!
Currently, this appears to stop roof regions being generated above air but has one known
flaw in that if you have a single solid layer above a hole then the region over the hole
will not be roof but everywhere else that is exposed to air above will be roof. Not sure
if this is easily solvable. However, if there are two solid layers above the hole then the
top layer will be roof over the hole.