Add version of moveInside that works on single polygonrefs
Otherwise you have to pack it in a Polygons and such and that's just overhead.
I'm not happy with this one because it duplicates the code, but I'm not seeing an easy solution without spending a lot of time on that trivial matter. If this research gets merged we'll fix that up.

Contributes to issue CURA-4523.