Now limits how far points are moved to be inside the combing boundary.
The limit is 1mm which should be satisfactory for all but the weirdest of scenarios.

Before, when the distance was not limited, the points could be moved far away to be inside
the combing boundary when combing was set to "no skin" and that would play havoc with the
choosing of the shortest move. Now, if a couple of points are not moved because they are
both more than 1mm outside of the combing boundary then the travel distance between them
will not be penalised as long as the direct path between those points does not cross the
combing boundary.