Revert back to using the combing distance when direct path between two points hits the boundary.
The difference from before is that the combing boundary has been simplified which greatly
reduces the amount of processor time required to calculate the combing distance.

Also, the line grid that is used by the combing code is only now created when a combed
distance is required which saves time.