Fix avoiding collisions
There seems to be something strange going on with the moveOutside functionality... Maybe it's not meant to work like this. When I use the actual areas it needs to avoid and use the maximum movement speed as distance, it seems to work all right. However we can't use the precise diameter of the branch any more then so we must be pessimistic and avoid a bit too far.

Contributes to issue CURA-4523.