Make search for interesting triangles more efficient.
Not much point in generating two sets and then intersecting them, just make one set.

This could be made more efficient still because having found the triangles that are within
the thickest layer of interest, if we subsequently look at thinner layers we don't need to
search through all of the triangles in the mesh, only the ones that were within the last
layer we were checking.