Move position data within tree nodes
The main idea is that we now have one set of nodes per layer rather than a set of positions per layer and then a mapping from positions to nodes that contain more metadata.
This required a whole rewrite of the dropping-down part of the algorithm though. This rewrite fixed a bug with merging the nodes, so it's useful as well. It makes the code less error-prone because the set and the map no longer need to be kept in sync. It requires an implementation of hashing these nodes, only by their position to keep the positions unique. I also refactored inserting nodes into the set out into a separate function to prevent repetition.

Contributes to issue CURA-4523.