No longer delete nodes that will collide with the model
But we don't include them in the relaxation step either. They remain stationary. Next we'll want to move them towards the model, even, and remove them when they actually hit the model. And then we need to make them smaller to prepare for hitting the model.

Contributes to issue CURA-4523.