Split nodes per collision area
If a node is inside a polygon part of the collision area, it gets grouped together. These groups all get their own minimum spanning tree and all move closer to each other (but not to nodes of other groups). This causes nodes that fall within the collision area of an object to cluster together.
The original code to move further inside is commented out for now. I plan to reimplement that next but I wanted to save that code.

Contributes to issue CURA-4523.