Move nodes towards the inside again
If they can't reach the build plate, nodes should move towards the inside of the model so that they collide with the model as quickly as possible.
Now they first try to move towards each other. Then afterwards, they will move towards the inside of the model. This is two steps, so then the sum of these directions is normalised to the maximum step size and this is then taken as the actual step.

Contributes to issue CURA-4523.