Add bool arg to LayerPlan::addLinesByOptimizer() to enable travel optimization.
When true, the travel distances used in the optimizer are based on the combing distance
rather than the "as the crow flies" distance. As this can be time consuming for some models
and infill patterns/density/etc. it is now an option rather than being always enabled.

When printing infill, the value of "infill_enable_travel_optimization" controls whether
this is enabled or not.