Improve grouping of nodes based on parts
The parts that it was based on were the parts for node radius 0. The problem is that for different radii of nodes there are different parts, even a different number of parts, and so the groups may be different. So now we still use the parts for node radius 0 but take the closest part instead of requiring that it's inside a part.

Contributes to issue CURA-4523.