Prevent nodes from persisting through the body of the model
If they've hit the model and their area is completely reduced to 0, delete them so that they don't shine through on the bottom side.

This has a bit of a limitation that is similar to the limitation with support X/Y offset in that it may shine through on the bottom side if the model is too thin and diagonal. If the branch is so wide that it shines through on the bottom side while still appearing on the top side, it won't be deleted and will continue to be handled on the bottom side of the model. I consider that rare though and if it occurs it won't ruin your print, just attach a bit too well to the model's underside.

Contributes to issue CURA-4523.