Allow contact points falling between consecutive overhang areas
In a sloped surface there is a zone depending on the overhang angle that is not marked as overhang, because the previous layer supports it. Allow points that fall within half of the width of this zone nearby the overhang area to snap towards the overhang area, so that all the grid points that fall in between those overhang areas to be used as well. This improves the spread of the points.

Contributes to issue CURA-4523.