Allow further movements because of distance between radius samples
When switching to the next radius sample, there might be a jump in the avoidance area. So your branch needs to be able to make larger jumps otherwise it'll end up inside the avoidance area and suddenly change direction to go towards the model instead of towards the build plate.

Contributes to issue CURA-4523.