Don't insert temperature command if there is a zero-length move in an extruder plan
If there are only zero-length moves in an extruder plan, it's not properly detected that the extruder plan is empty. To properly detect that, you'd need to compute the total time estimate and see if that is zero. This takes computing power though, but luckily we already have the time estimate here just in time to do something about it.

Contributes to issue CURA-4814.