If the first layer is empty and the second is not empty, always remove the first layer.
The first layer can be empty when the slicing tolerance is not inclusive and support won't
be generated in such a small gap so move the model down.

This is done without consulting remove_empty_first_layers as that setting is hidden when
support is in use and so it could be true or not and the user would not know so I think
it's better to ignore it here.