Also reset E using G92 on BFB and Makerbot flavours
The original commit says that it's not supported well. However, the RepRapWiki says it's supported nowadays. Maybe the earlier firmware had some trouble there.
I'm still not resetting the E value after E10,000 though, just to be sure. It's only resetting the E value in an extruder switch, which was causing a crash because it assumed (and asserted) that the E value was 0 after an extruder switch.

Contributes to issue CURA-4918.