Fix a number of compilation errors
It compiles and runs now, but the code has taken a few hits. I'll try to resolve those one by one in the following commits.
One thing that needed to happen was to move all implementation into the header file for UnionFind.h since it is a class template.

Contributes to issue CURA-4017.