Store the crossings encountered while generating line infill
The data structure is precarious right now, because if the number of vertices changes or the number of polygons changes, or even if the polygons change their ordering, this will be some number of vertices short or even access uninitialised memory. We need to have the polygons of the actual infill area fixed before the rotation is applied to it. I've put that in TODO comments for now.

Contributes to issue CURA-4017.