Do matching of joining on previous crossing, but match to segment
Because we introduce new segments while connecting two lines, those new segments will obviously not be in the UnionFind. We could either add them there, introducing additional data to that data structure, or we could keep track of the previous actual infill line that we encountered. I opt for the second option. previous_crossing is the latest actual crossing of an infill line with the polygon, and previous_segment is the latest segment that we've been drawing, which may be an actual infill line but may also be just a connecting piece.

Contributes to issue CURA-4017.